Fix backend panel selects
