+++ src/qwen_tui/tui/__init__.py
@@ -1,6 +1,8 @@
 """TUI components for Qwen-TUI."""
 
-from .app import QwenTUIApp, InputPanel, ThinkingWidget, ActionWidget
+from .app import QwenTUIApp
+from .input_panel import InputPanel
+from .widgets import ThinkingWidget, ActionWidget
 from .backend_panel import BackendPanel
 from .chat_panel import ChatPanel
